[crt][fuzzer] Allow variable-length TORCW mutations
Currently, the two string arguments to strcmp calls are truncated to the
length of the shorter of the two before they are inserted into the table
of recent compares.

With this commit, the two strings are recorded without truncation (up to
a maximum length of Word::GetMaxSize()) and any occurence of either of
the two can be replaced by the other via the TORCW-based mutator. To
make this possible, the concept of a position hint is extended to not
only contain the buffer position at which the string to be replaced has
been found in the input, but also the length of the string.